backend/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
sentence-transformers
groq
langchain
langchain-groq
python-multipart
